AI_analysis_engine/requirements.txt
pymongo==4.6.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
llama-cpp-python==0.2.11
paddlepaddle==2.5.1
paddleocr==2.7.0.3
PyMuPDF==1.23.8